# eventreg/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Launch-proof event registration. Handles the \"everyone hits register at once\"\nstampede: never double-books a seat, auto-waitlists when full, and promotes\nfrom the queue when someone cancels.\n\nAuth: call /auth/register or /auth/login to get a JWT, then click\n**Authorize** above and paste `Bearer \u003ctoken\u003e`.",
        "title": "Event Registration API",
        "contact": {
            "name": "Aditya Choudhary"
        },
        "version": "1.0"
    },
    "host": "localhost:8082",
    "basePath": "/",
    "paths": {
...
        "/events": {
            "get": {
                "description": "Public. Returns a page envelope: {items, total, limit, offset, has_more}.\n`limit` defaults to 20 and is clamped to a maximum of 100.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "List events (paginated)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "page size (default 20, max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "rows to skip (default 0)",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.EventPage"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Requires authentication. `id` in the body is ignored — the database assigns it.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Create an event",
                "parameters": [
                    {
                        "description": "name and seats",
                        "name": "event",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Event"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Event"
                        }
                    },
                    "400": {
                        "description": "malformed JSON",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "missing or invalid token",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "name empty or seats \u003c= 0",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/events/stats": {
            "get": {
                "description": "Public. Booking count and seats sold per event, via GROUP BY with a\nLEFT JOIN — so events with zero bookings still appear.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Per-event sales stats",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.EventStat"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/events/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "events"
                ],
                "summary": "Get one event",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "event id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Event"
                        }
                    },
                    "400": {
                        "description": "id is not a number",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "no such event",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/events/{id}/book": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Requires authentication. The booking is attributed to the token's user.\n\n**Idempotent**: repeating the same request returns the original booking\nwith `outcome: already_booked` and does not consume more seats.\n\nStatus codes carry the outcome:\n- **201** `booked` — a new booking exists\n- **202** `waitlisted` — event was full, you were queued\n- **200** `already_booked` / `already_waitlisted` — replay, nothing changed",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookings"
                ],
                "summary": "Book seats (auto-waitlists if full)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "event id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "seats to book",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.BookRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "idempotent replay",
                        "schema": {
                            "$ref": "#/definitions/models.BookingResult"
                        }
                    },
                    "201": {
                        "description": "booked",
                        "schema": {
                            "$ref": "#/definitions/models.BookingResult"
                        }
                    },
                    "202": {
                        "description": "waitlisted",
                        "schema": {
                            "$ref": "#/definitions/models.BookingResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "no such event",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "422": {
                        "description": "seats \u003c= 0",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Requires authentication. Frees your seats and immediately promotes the\nlongest-waiting users who fit — all in one transaction, so the freed seats\ncannot be taken by a new arrival before the queue gets them.\nThe response lists who was promoted.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bookings"
                ],
                "summary": "Cancel a booking (promotes from the waitlist)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "event id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CancelResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "you have no booking for this event",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/me/bookings": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Requires authentication. Scoped to the token's user — there is no id in the\nURL to tamper with, so you cannot read anyone else's bookings.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "me"
                ],
                "summary": "My bookings",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.BookingWithEvent"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/me/waitlist": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Requires authentication. `position` is 1-based and computed at read time\nfrom arrival order — it is not stored, so it is never stale.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "me"
                ],
                "summary": "My waitlist entries (with queue position)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.WaitlistEntryWithEvent"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.BookRequest": {
            "type": "object",
            "properties": {
                "seats": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
...
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "event not found"
                }
            }
        },
...
        "models.Booking": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "event_id": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "idempotent": {
                    "description": "Idempotent tells the caller \"this booking already existed; your request\nwas a duplicate and changed nothing.\" Not persisted — it's a property of\nTHIS response, not of the row. Lets the handler answer 200 (here it is\nagain) instead of 201 (something new was made).",
                    "type": "boolean"
                },
                "seats": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.BookingResult": {
            "type": "object",
            "properties": {
                "booking": {
                    "$ref": "#/definitions/models.Booking"
                },
                "outcome": {
                    "type": "string"
                },
                "waitlist": {
                    "$ref": "#/definitions/models.WaitlistEntry"
                }
            }
        },
        "models.BookingWithEvent": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "event_id": {
                    "type": "integer"
                },
                "event_name": {
                    "type": "string"
                },
                "event_seats_left": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "idempotent": {
                    "description": "Idempotent tells the caller \"this booking already existed; your request\nwas a duplicate and changed nothing.\" Not persisted — it's a property of\nTHIS response, not of the row. Lets the handler answer 200 (here it is\nagain) instead of 201 (something new was made).",
                    "type": "boolean"
                },
                "seats": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "models.CancelResult": {
            "type": "object",
            "properties": {
                "event_id": {
                    "type": "integer"
                },
                "promoted": {
                    "description": "waitlisted users who just got in",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Booking"
                    }
                },
                "seats_freed": {
                    "type": "integer"
                },
                "seats_left": {
                    "type": "integer"
                }
            }
        },
        "models.Event": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "seats": {
                    "type": "integer"
                }
            }
        },
        "models.EventPage": {
            "type": "object",
            "properties": {
                "has_more": {
                    "description": "HasMore saves the client doing offset+limit \u003c total arithmetic, and lets\nus change the underlying strategy (e.g. to cursors) without breaking them.",
                    "type": "boolean"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Event"
                    }
                },
                "limit": {
                    "description": "page size actually applied (may be clamped)",
                    "type": "integer"
                },
                "offset": {
                    "description": "rows skipped",
                    "type": "integer"
                },
                "total": {
                    "description": "total matching rows, ignoring limit/offset",
                    "type": "integer"
                }
            }
        },
        "models.EventStat": {
            "type": "object",
            "properties": {
                "booking_count": {
                    "description": "COUNT(*) — how many bookings",
                    "type": "integer"
                },
                "event_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "seats_booked": {
                    "description": "SUM(seats) — how many seats sold",
                    "type": "integer"
                },
                "seats_left": {
                    "type": "integer"
                }
            }
        },
...
        "models.WaitlistEntryWithEvent": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "event_id": {
                    "type": "integer"
                },
                "event_name": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "position": {
                    "description": "Position is COMPUTED at read time (1-based), never stored. Storing a\nposition column would mean renumbering every row behind someone who\nleaves the queue — an expensive write amplification and a fresh source\nof races. created_at already encodes the order, so we derive position\nfrom it instead. Prefer deriving over duplicating.",
                    "type": "integer"
                },
                "seats": {
                    "type": "integer"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "JWT from /auth/login. Format: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}